// src/Interfaces/MovieDetails.tsx
export interface MovieDetails{
    "Title": string;
    "Year": string;
    "Rated": string;
    "Released": string;
    "Runtime": string;
    "Genre": string;
    "Director": string;
    "Writer": string;
    "Actors": string;
    "Plot": string;
    "Language": string;
    "Country": string;
    "Awards": string;
    "Poster": string;
}

export const movieExample:MovieDetails = {
    "Title": "string",
    "Year": "string",
    "Rated": "string",
    "Released": "string",
    "Runtime": "string",
    "Genre": "string",
    "Director": "string",
    "Writer": "string",
    "Actors": "string",
    "Plot": "string",
    "Language": "string",
    "Country": "string",
    "Awards": "string",
    "Poster": "string"
}